Rearranging file structure
