+++ automotive/src/main/java/com/ixam97/carStatsViewer/activities/MainActivity.kt
@@ -1,8 +1,7 @@
 package com.ixam97.carStatsViewer.activities
 
 import com.ixam97.carStatsViewer.*
-import com.ixam97.carStatsViewer.objects.*
-import com.ixam97.carStatsViewer.services.*
+import com.ixam97.carStatsViewer.dataManager.*
 import android.app.Activity
 import android.app.PendingIntent
 import android.car.VehicleGear
@@ -19,9 +18,11 @@ import android.os.Looper
 import android.os.SystemClock
 import android.view.View
 import android.widget.Toast
+import com.ixam97.carStatsViewer.appPreferences.AppPreferences
 import com.ixam97.carStatsViewer.plot.enums.*
 import com.ixam97.carStatsViewer.plot.graphics.PlotPaint
 import com.ixam97.carStatsViewer.views.PlotView
+import com.ixam97.carStatsViewer.dataManager.DataManagers
 import kotlinx.android.synthetic.main.activity_main.*
 import java.util.*
 import java.util.concurrent.TimeUnit
@@ -43,6 +44,8 @@ class MainActivity : Activity() {
     private lateinit var starterIntent: Intent
     private lateinit var context: Context
     private lateinit var appPreferences: AppPreferences
+
+    private var selectedDataManager = DataManagers.CURRENT_TRIP.dataManager
 
     private var updateUi = false
     private var lastPlotUpdate: Long = 0L
@@ -71,13 +74,13 @@ class MainActivity : Activity() {
         InAppLogger.log("MainActivity.onResume")
 
         if (appPreferences.consumptionUnit) {
-            DataCollector.CurrentTripDataManager.consumptionPlotLine.Configuration.Unit = "Wh/km"
-            DataCollector.CurrentTripDataManager.consumptionPlotLine.Configuration.LabelFormat = PlotLineLabelFormat.NUMBER
-            DataCollector.CurrentTripDataManager.consumptionPlotLine.Configuration.Divider = 1f
+            selectedDataManager.consumptionPlotLine.Configuration.Unit = "Wh/km"
+            selectedDataManager.consumptionPlotLine.Configuration.LabelFormat = PlotLineLabelFormat.NUMBER
+            selectedDataManager.consumptionPlotLine.Configuration.Divider = 1f
         } else {
-            DataCollector.CurrentTripDataManager.consumptionPlotLine.Configuration.Unit = "kWh/100km"
-            DataCollector.CurrentTripDataManager.consumptionPlotLine.Configuration.LabelFormat = PlotLineLabelFormat.FLOAT
-            DataCollector.CurrentTripDataManager.consumptionPlotLine.Configuration.Divider = 10f
+            selectedDataManager.consumptionPlotLine.Configuration.Unit = "kWh/100km"
+            selectedDataManager.consumptionPlotLine.Configuration.LabelFormat = PlotLineLabelFormat.FLOAT
+            selectedDataManager.consumptionPlotLine.Configuration.Divider = 10f
         }
 
         main_power_gage.maxValue = if (appPreferences.consumptionPlotSingleMotor) 170f else 300f
@@ -98,15 +101,14 @@ class MainActivity : Activity() {
             main_consumption_plot.secondaryDimension != null -> getString(R.string.main_button_hide_speed)
             else -> getString(R.string.main_button_show_speed)
         }
-
-        DataCollector.CurrentTripDataManager.consumptionPlotLine.plotPaint = PlotPaint.byColor(getColor(R.color.primary_plot_color), PlotView.textSize)
-        DataCollector.CurrentTripDataManager.consumptionPlotLine.secondaryPlotPaint = when {
+        selectedDataManager.consumptionPlotLine.plotPaint = PlotPaint.byColor(getColor(R.color.primary_plot_color), PlotView.textSize)
+        selectedDataManager.consumptionPlotLine.secondaryPlotPaint = when {
             appPreferences.consumptionPlotSecondaryColor -> PlotPaint.byColor(getColor(R.color.secondary_plot_color_alt), PlotView.textSize)
             else -> PlotPaint.byColor(getColor(R.color.secondary_plot_color), PlotView.textSize)
         }
 
-        DataCollector.CurrentTripDataManager.chargePlotLine.plotPaint = PlotPaint.byColor(getColor(R.color.charge_plot_color), PlotView.textSize)
-        DataCollector.CurrentTripDataManager.chargePlotLine.secondaryPlotPaint = when {
+        selectedDataManager.chargePlotLine.plotPaint = PlotPaint.byColor(getColor(R.color.charge_plot_color), PlotView.textSize)
+        selectedDataManager.chargePlotLine.secondaryPlotPaint = when {
             appPreferences.chargePlotSecondaryColor -> PlotPaint.byColor(getColor(R.color.secondary_plot_color_alt), PlotView.textSize)
             else -> PlotPaint.byColor(getColor(R.color.secondary_plot_color), PlotView.textSize)
         }
@@ -127,7 +129,7 @@ class MainActivity : Activity() {
 
         startForegroundService(Intent(this, DataCollector::class.java))
 
-        mainActivityPendingIntent = PendingIntent.getActivity(
+        DataCollector.mainActivityPendingIntent = PendingIntent.getActivity(
             this, 0, intent, PendingIntent.FLAG_IMMUTABLE
         )
 
@@ -175,24 +177,24 @@ class MainActivity : Activity() {
         updateGages()
 
         main_gage_avg_consumption_text_view.text = "  Ø %s".format(getAvgConsumptionString())
-        main_gage_distance_text_view.text = "  %s".format(getTraveledDistanceString(DataCollector.CurrentTripDataManager.traveledDistance))
-        main_gage_used_power_text_view.text = "  %s".format(getEnergyString(DataCollector.CurrentTripDataManager.usedEnergy))
-        main_gage_time_text_view.text = "  %s".format(getElapsedTimeString(DataCollector.CurrentTripDataManager.travelTime))
-        main_gage_charged_energy_text_view.text = "  %s".format(getEnergyString(DataCollector.CurrentTripDataManager.chargedEnergy))
-        main_gage_charge_time_text_view.text = "  %s".format(getElapsedTimeString(DataCollector.CurrentTripDataManager.chargeTime))
+        main_gage_distance_text_view.text = "  %s".format(getTraveledDistanceString(selectedDataManager.traveledDistance))
+        main_gage_used_power_text_view.text = "  %s".format(getEnergyString(selectedDataManager.usedEnergy))
+        main_gage_time_text_view.text = "  %s".format(getElapsedTimeString(selectedDataManager.travelTime))
+        main_gage_charged_energy_text_view.text = "  %s".format(getEnergyString(selectedDataManager.chargedEnergy))
+        main_gage_charge_time_text_view.text = "  %s".format(getElapsedTimeString(selectedDataManager.chargeTime))
     }
 
     private fun getCurrentSpeedString(): String {
-        return "${(DataCollector.CurrentTripDataManager.currentSpeed * 3.6).toInt()} km/h"
+        return "${(selectedDataManager.currentSpeed * 3.6).toInt()} km/h"
     }
 
     private fun getAvgSpeedString(): String {
         return " %d km/h".format(
-            ((DataCollector.CurrentTripDataManager.traveledDistance / 1000) / (DataCollector.CurrentTripDataManager.travelTime.toFloat() / 3_600_000)).toInt())
+            ((selectedDataManager.traveledDistance / 1000) / (selectedDataManager.travelTime.toFloat() / 3_600_000)).toInt())
     }
 
     private fun getCurrentPowerString(detailed : Boolean = true): String {
-        val rawPower = DataCollector.CurrentTripDataManager.currentPower / 1_000_000
+        val rawPower = selectedDataManager.currentPower / 1_000_000
 
         return when {
             !detailed && rawPower.absoluteValue >= 10 -> "%d kW".format(Locale.ENGLISH, rawPower.roundToInt())
@@ -210,29 +212,29 @@ class MainActivity : Activity() {
     }
 
     private fun getInstConsumptionString(): String {
-        if (DataCollector.CurrentTripDataManager.currentSpeed <= 0) {
+        if (selectedDataManager.currentSpeed <= 0) {
             return "-/-"
         }
         if (!appPreferences.consumptionUnit) {
             return "%.1f kWh/100km".format(
                 Locale.ENGLISH,
-                ((DataCollector.CurrentTripDataManager.currentPower / 1000) / (DataCollector.CurrentTripDataManager.currentSpeed * 3.6))/10)
+                ((selectedDataManager.currentPower / 1000) / (selectedDataManager.currentSpeed * 3.6))/10)
         }
-        return "${((DataCollector.CurrentTripDataManager.currentPower / 1000) / (DataCollector.CurrentTripDataManager.currentSpeed * 3.6)).toInt()} Wh/km"
+        return "${((selectedDataManager.currentPower / 1000) / (selectedDataManager.currentSpeed * 3.6)).toInt()} Wh/km"
     }
 
     private fun getAvgConsumptionString(): String {
         val unitString = if (appPreferences.consumptionUnit) "Wh/km" else "kWh/100km"
-        if (DataCollector.CurrentTripDataManager.traveledDistance <= 0) {
+        if (selectedDataManager.traveledDistance <= 0) {
             return "-/- $unitString"
         }
         if (!appPreferences.consumptionUnit) {
             return "%.1f %s".format(
                 Locale.ENGLISH,
-                (DataCollector.CurrentTripDataManager.usedEnergy /(DataCollector.CurrentTripDataManager.traveledDistance /1000))/10,
+                (selectedDataManager.usedEnergy /(selectedDataManager.traveledDistance /1000))/10,
                 unitString)
         }
-        return "${(DataCollector.CurrentTripDataManager.usedEnergy /(DataCollector.CurrentTripDataManager.traveledDistance /1000)).toInt()} $unitString"
+        return "${(selectedDataManager.usedEnergy /(selectedDataManager.traveledDistance /1000)).toInt()} $unitString"
     }
 
     private fun getTraveledDistanceString(traveledDistance: Float): String {
@@ -241,24 +243,24 @@ class MainActivity : Activity() {
 
     private fun setUiVisibilities() {
 
-        if (main_button_dismiss_charge_plot.isEnabled == DataCollector.CurrentTripDataManager.chargePortConnected)
-            main_button_dismiss_charge_plot.isEnabled = !DataCollector.CurrentTripDataManager.chargePortConnected
+        if (main_button_dismiss_charge_plot.isEnabled == selectedDataManager.chargePortConnected)
+            main_button_dismiss_charge_plot.isEnabled = !selectedDataManager.chargePortConnected
 
-        if (main_charge_layout.visibility == View.GONE && DataCollector.CurrentTripDataManager.chargePortConnected) {
+        if (main_charge_layout.visibility == View.GONE && selectedDataManager.chargePortConnected) {
             main_consumption_layout.visibility = View.GONE
             main_charge_layout.visibility = View.VISIBLE
         }
     }
 
     private fun updateGages() {
-        if ((DataCollector.CurrentTripDataManager.currentPower / 1_000_000).absoluteValue > 10 && true) { // Add Setting!
-            val newValue = (DataCollector.CurrentTripDataManager.currentPower / 1_000_000).toInt()
+        if ((selectedDataManager.currentPower / 1_000_000).absoluteValue > 10 && true) { // Add Setting!
+            val newValue = (selectedDataManager.currentPower / 1_000_000).toInt()
             main_power_gage.setValue(newValue)
         } else {
-            main_power_gage.setValue(DataCollector.CurrentTripDataManager.currentPower / 1_000_000)
+            main_power_gage.setValue(selectedDataManager.currentPower / 1_000_000)
         }
-        main_charge_gage.setValue(-DataCollector.CurrentTripDataManager.currentPower / 1_000_000)
-        main_SoC_gage.setValue((100f / DataCollector.CurrentTripDataManager.maxBatteryLevel * DataCollector.CurrentTripDataManager.batteryLevel).roundToInt())
+        main_charge_gage.setValue(-selectedDataManager.currentPower / 1_000_000)
+        main_SoC_gage.setValue((100f / selectedDataManager.maxBatteryLevel * selectedDataManager.batteryLevel).roundToInt())
 
         var consumptionValue: Float? = null
 
@@ -266,8 +268,8 @@ class MainActivity : Activity() {
             main_consumption_gage.gageUnit = "Wh/km"
             main_consumption_gage.minValue = -300f
             main_consumption_gage.maxValue = 600f
-            if (DataCollector.CurrentTripDataManager.currentSpeed * 3.6 > 3) {
-                main_consumption_gage.setValue(((DataCollector.CurrentTripDataManager.currentPower / 1000) / (DataCollector.CurrentTripDataManager.currentSpeed * 3.6)).toInt())
+            if (selectedDataManager.currentSpeed * 3.6 > 3) {
+                main_consumption_gage.setValue(((selectedDataManager.currentPower / 1000) / (selectedDataManager.currentSpeed * 3.6)).toInt())
             } else {
                 main_consumption_gage.setValue(consumptionValue)
             }
@@ -276,8 +278,8 @@ class MainActivity : Activity() {
             main_consumption_gage.gageUnit = "kWh/100km"
             main_consumption_gage.minValue = -30f
             main_consumption_gage.maxValue = 60f
-            if (DataCollector.CurrentTripDataManager.currentSpeed * 3.6 > 3) {
-                main_consumption_gage.setValue(((DataCollector.CurrentTripDataManager.currentPower / 1000) / (DataCollector.CurrentTripDataManager.currentSpeed * 3.6f))/10)
+            if (selectedDataManager.currentSpeed * 3.6 > 3) {
+                main_consumption_gage.setValue(((selectedDataManager.currentPower / 1000) / (selectedDataManager.currentSpeed * 3.6f))/10)
             } else {
                 main_consumption_gage.setValue(consumptionValue)
             }
@@ -287,7 +289,7 @@ class MainActivity : Activity() {
     private fun updatePlots(){
         // if (appPreferences.plotDistance == 3) main_consumption_plot.dimensionRestriction = dimensionRestrictionById(appPreferences.plotDistance)
 
-        main_charge_plot.dimensionRestriction = TimeUnit.MINUTES.toNanos((TimeUnit.MILLISECONDS.toMinutes(DataCollector.CurrentTripDataManager.chargeTime) / 5) + 1) * 5 + TimeUnit.MILLISECONDS.toNanos(1)
+        main_charge_plot.dimensionRestriction = TimeUnit.MINUTES.toNanos((TimeUnit.MILLISECONDS.toMinutes(selectedDataManager.chargeTime) / 5) + 1) * 5 + TimeUnit.MILLISECONDS.toNanos(1)
 
         if (SystemClock.elapsedRealtime() - lastPlotUpdate > 1_000L) {
             if (main_consumption_layout.visibility == View.VISIBLE) {
@@ -322,7 +324,7 @@ class MainActivity : Activity() {
         finish()
         startActivity(intent)
         InAppLogger.log("MainActivity.resetStats")
-        DataCollector.CurrentTripDataManager.reset()
+        selectedDataManager.reset()
         sendBroadcast(Intent(getString(R.string.save_trip_data_broadcast)))
     }
 
@@ -338,7 +340,7 @@ class MainActivity : Activity() {
         main_radio_group_distance.check(plotDistanceId)
 
         main_consumption_plot.reset()
-        main_consumption_plot.addPlotLine(DataCollector.CurrentTripDataManager.consumptionPlotLine)
+        main_consumption_plot.addPlotLine(selectedDataManager.consumptionPlotLine)
         // main_consumption_plot.setPlotMarkers(DataManager.plotMarkers)
         // main_consumption_plot.visibleMarkerTypes.add(PlotMarkerType.CHARGE)
         // main_consumption_plot.visibleMarkerTypes.add(PlotMarkerType.PARK)
@@ -348,7 +350,7 @@ class MainActivity : Activity() {
             else -> getString(R.string.main_button_show_speed)
         }
 
-        DataCollector.CurrentTripDataManager.consumptionPlotLine.secondaryPlotPaint = when {
+        selectedDataManager.consumptionPlotLine.secondaryPlotPaint = when {
             appPreferences.consumptionPlotSecondaryColor -> PlotPaint.byColor(getColor(R.color.secondary_plot_color_alt), PlotView.textSize)
             else -> PlotPaint.byColor(getColor(R.color.secondary_plot_color), PlotView.textSize)
         }
@@ -366,9 +368,9 @@ class MainActivity : Activity() {
         main_consumption_plot.invalidate()
 
         main_charge_plot.reset()
-        main_charge_plot.addPlotLine(DataCollector.CurrentTripDataManager.chargePlotLine)
+        main_charge_plot.addPlotLine(selectedDataManager.chargePlotLine)
 
-        DataCollector.CurrentTripDataManager.chargePlotLine.secondaryPlotPaint = when {
+        selectedDataManager.chargePlotLine.secondaryPlotPaint = when {
             appPreferences.chargePlotSecondaryColor -> PlotPaint.byColor(getColor(R.color.secondary_plot_color_alt), PlotView.textSize)
             else -> PlotPaint.byColor(getColor(R.color.secondary_plot_color), PlotView.textSize)
         }
@@ -419,7 +421,7 @@ class MainActivity : Activity() {
                         VehiclePropertyIds.GEAR_SELECTION
                     )
                     putExtra(EmulatorIntentExtras.TYPE, EmulatorIntentExtras.TYPE_INT)
-                    if (DataCollector.CurrentTripDataManager.currentGear == VehicleGear.GEAR_PARK) {
+                    if (selectedDataManager.currentGear == VehicleGear.GEAR_PARK) {
                         putExtra(
                             EmulatorIntentExtras.VALUE,
                             VehicleGear.GEAR_DRIVE
