+++ automotive/src/main/java/com/ixam97/carStatsViewer/appPreferences/AppPreference.kt
@@ -1,0 +1,17 @@
+package com.ixam97.carStatsViewer.appPreferences
+
+enum class AppPreference {
+    DEBUG,
+    NOTIFICATIONS,
+    CONSUMPTION_UNIT,
+    EXPERIMENTAL_LAYOUT,
+    DEEP_LOG,
+    PLOT_SHOW_SPEED,
+    PLOT_DISTANCE,
+    CONSUMPTION_PLOT_SINGLE_MOTOR,
+    CONSUMPTION_PLOT_SECONDARY_COLOR,
+    CONSUMPTION_PLOT_VISIBLE_GAGES,
+    CHARGE_PLOT_SECONDARY_COLOR,
+    CHARGE_PLOT_VISIBLE_GAGES,
+    CHARGE_PLOT_DIMENSION
+}
