--- automotive/src/main/java/com/ixam97/carStatsViewer/objects/ChargeCurve.kt
@@ -1,9 +1,0 @@
-package com.ixam97.carStatsViewer.objects
-
-import com.ixam97.carStatsViewer.plot.objects.PlotLineItem
-
-data class ChargeCurve(
-    var chargePlotLine: List<PlotLineItem>,
-    var chargeTime: Long,
-    var chargedEnergy: Float
-) {}
